====== Complex ====== Finished the (currently untested) Complex.java file. This allows for a representation of Complex data using double values. The test-file is also added, and has not been completed. TODO: Finish testing all methods to ensure perfect functionality of the class.
======
Fourier Transform
======
The FourierData.java file contains ArrayList-s which hold the
frequency-domain data for the FFT. It is supposed to include checking
for upper and lower bounds on frequencies (i.e. if your time-domain data
is from [-n/2,n/2] instead of [0,n], then the frequencies go from
[-f/2,f/2] instead of [0,f] where f is the sample rate). However, this
may not be necessary.
TODO: finalize the constructor for the object and all private methods,
and determine if it needs any public methods.

The FourierTransform.java file gives the functionality of the FFT
(hereby labelled the DDFT, or the DiscreteTime -> DiscreteFrequency
Fourier Transform). I am creating a method that calculates the
FourierTransform of Complex[] and double[] data only, as it is all I
will need for my purposes. This FFT is created from the methods
described in Queen's University's MTHE 335 course, and from Princeton's
intro to CS FFT java file.
TODO: complete the implementation of the FFT as given by princeton, and
implement the Inverse FFT functionality.