Created the TimeData object, which will hold all the Sound components of the project. In the FourierTransform file, generated a method for converting the Princeton FFT code into a FourierData object. Fleshed out the TimeData and FourierData objects, and added some (unfinished) test code for each of them.
TODO
Realized that the ArrayLists for the TimeData and FourierData classes
could better be compressed by a HashMap, so I will spend some time
converting those before ensuring test cases work well.
Need to finish testing the FourierTransform method. Once it is working,
we can begin the inverse fourier transform method, and we can start
working on the Hilbertizer.